+++ lib/transformers.ts
@@ -20,37 +20,40 @@ export interface MapCallback<T, U> {
 
 /**
  * Lazily calls a defined callback function for each element of an iterable, and
- * returns a new iterator of the results.
+ * returns a new iterable of the results.
  * @param it - The iterable being mapped.
  * @param {MapCallback} f - A function that accepts up to three arguments. The
  * map method calls `f` function one time for each item in the iterable.
  * @typeParam T - Type of items in `it`.
  * @typeParam U - Return type of `f`.
- * @returns An iterator of `f` applied to items of `it`.
+ * @returns An iterable of `f` applied to items of `it`.
  */
-export function* map<T, U = T>(
+export function map<T, U = T>(
   it: Iterable<T>,
   f: MapCallback<T, U>,
-): IterableIterator<U> {
-  let index = 0;
-  for (const item of it) {
-    yield f(item, index, stripIterable(it));
-    index++;
-  }
+): Iterable<U> {
+  return {
+    *[Symbol.iterator](): IterableIterator<U> {
+      let index = 0;
+      for (const item of it) yield f(item, index++, stripIterable(it));
+    },
+  };
 }
 
 /**
  * Returns a new iterable containing the first `n` items of `it`.
  * @param it - The iterable being taken from.
  * @param n - The number of items to take.
- * @typeParam T - The type of items in both `it` and the returned iterator.
- * @returns A new iterator of `it` which terminates after `n` items.
+ * @typeParam T - The type of items in both `it` and the returned iterable.
+ * @returns A new iterable of `it` which terminates after `n` items.
  */
-export function* take<T>(it: Iterable<T>, n: number): IterableIterator<T> {
-  const iterator = it[Symbol.iterator]();
-  for (let i = 0; i < n; i++) {
-    yield iterator.next().value;
-  }
+export function take<T>(it: Iterable<T>, n: number): Iterable<T> {
+  return {
+    *[Symbol.iterator](): IterableIterator<T> {
+      const iterator = it[Symbol.iterator]();
+      for (let i = 0; i < n; i++) yield iterator.next().value;
+    },
+  };
 }
 
 /**
@@ -61,27 +64,30 @@ export function* take<T>(it: Iterable<T>, n: number): IterableIterator<T> {
  * arguments. The cut method calls `f` one time for each item in the iterable.
  * @param includeLast - Whether the item for which `f` returns true should be
  * included.
- * @typeParam T - The type of items in both `it` and the returned iterator.
- * @returns A new iterator of `it` which terminates
+ * @typeParam T - The type of items in both `it` and the returned iterable.
+ * @returns A new iterables of `it` which terminates
  * @alias until
  */
-export function* until<T>(
+export function until<T>(
   it: Iterable<T>,
   f: IterablePredicateCallback<T>,
   includeLast = true,
-): IterableIterator<T> {
-  let index = 0;
-  for (const item of it) {
-    const done = f(item, index, stripIterable(it));
-    if (done) {
-      if (includeLast) {
+): Iterable<T> {
+  return {
+    *[Symbol.iterator](): IterableIterator<T> {
+      let index = 0;
+      for (const item of it) {
+        const done = f(item, index++, stripIterable(it));
+        if (done) {
+          if (includeLast) {
+            yield item;
+          }
+          break;
+        }
         yield item;
       }
-      break;
-    }
-    yield item;
-    index++;
-  }
+    },
+  };
 }
 
 /**
@@ -94,17 +100,20 @@ export function* until<T>(
  * @typeParam T - The type of items in `it`.
  * @returns A new iterable
  */
-export function* filter<T>(
+export function filter<T>(
   it: Iterable<T>,
   predicate: IterablePredicateCallback<T>,
-): IterableIterator<T> {
-  let index = 0;
-  for (const item of it) {
-    if (predicate(item, index, stripIterable(it))) {
-      yield item;
-    }
-    index++;
-  }
+): Iterable<T> {
+  return {
+    *[Symbol.iterator](): IterableIterator<T> {
+      let index = 0;
+      for (const item of it) {
+        if (predicate(item, index++, stripIterable(it))) {
+          yield item;
+        }
+      }
+    },
+  };
 }
 
 /**
@@ -112,14 +121,17 @@ export function* filter<T>(
  * {@link https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/entries | `Array.prototype.entries`}.
  * @param it - The iterable being indexed.
  * @typeParam T - The type of items in `it`.
- * @returns An iterator over pairs of indices and the items in `it`.
+ * @returns An iterable over pairs of indices and the items in `it`.
  */
-export function* indexedPairs<T>(it: Iterable<T>): Iterable<[number, T]> {
-  let index = 0;
-  for (const item of it) {
-    yield [index, item];
-    index++;
-  }
+export function indexedPairs<T>(it: Iterable<T>): Iterable<[number, T]> {
+  return {
+    *[Symbol.iterator](): IterableIterator<[number, T]> {
+      let index = 0;
+      for (const item of it) {
+        yield [index++, item];
+      }
+    },
+  };
 }
 
 /**
@@ -128,37 +140,38 @@ export function* indexedPairs<T>(it: Iterable<T>): Iterable<[number, T]> {
  * @param it - The iterable being chunkified.
  * @param chunkSize - The size of each chunk.
  * @typeParam T - The type of items in `it`.
- * @returns A new iterator over chunk arrays.
+ * @returns A new iterable over chunk arrays.
  */
-export function* chunkify<T>(
-  it: Iterable<T>,
-  chunkSize: number,
-): IterableIterator<T[]> {
+export function chunkify<T>(it: Iterable<T>, chunkSize: number): Iterable<T[]> {
   if (!(Number.isSafeInteger(chunkSize) && chunkSize > 0)) {
     throw new RangeError(
       `Expected \`chunkSize\` to be an integer from 1 and up, got \`${chunkSize}\``,
     );
   }
 
-  if (Array.isArray(it)) {
-    for (let index = 0; index < it.length; index += chunkSize) {
-      yield it.slice(index, index + chunkSize);
-      return;
-    }
-  }
+  return {
+    *[Symbol.iterator](): IterableIterator<T[]> {
+      if (Array.isArray(it)) {
+        for (let index = 0; index < it.length; index += chunkSize) {
+          yield it.slice(index, index + chunkSize);
+          return;
+        }
+      }
 
-  let chunk = [];
+      let chunk = [];
 
-  for (const value of it) {
-    chunk.push(value);
+      for (const value of it) {
+        chunk.push(value);
 
-    if (chunk.length === chunkSize) {
-      yield chunk;
-      chunk = [];
-    }
-  }
+        if (chunk.length === chunkSize) {
+          yield chunk;
+          chunk = [];
+        }
+      }
 
-  if (chunk.length > 0) {
-    yield chunk;
-  }
+      if (chunk.length > 0) {
+        yield chunk;
+      }
+    },
+  };
 }